# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Plataforma de Atendimento Omnichannel (SaaS Multi-Tenant) API",
    "description": "Especificação completa da API Go (Fiber + PostgreSQL RLS + Redis) estilo Chatwoot + Kommo + Flow Engine",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/conversations": {
      "get": {
        "summary": "Lista conversas ativas por status e atendente",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Lista de conversas" } }
      }
    },
...
    "/reports/conversations/export/csv": {
      "get": {
        "summary": "Download de relatório em formato CSV",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Arquivo CSV baixado" } }
      }
    },
...
    "/reports/conversations/export/pdf": {
      "get": {
        "summary": "Download PDF de conversas",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "PDF gerado" } }
      }
    },
...
  }
}